requirements.txt
mojito2
pyupbit
pykrx
schedule
requests
PyQt5
pandas_market_calendars
pandas
numpy
pytz
yfinance
lxml